requirements.txt
torch==2.4.0
torchvision
omegaconf
transformers
timm
open_clip_torch
einops
scipy
pillow
accelerate
gdown
huggingface-hub
einops
wandb
torch-fidelity
torchinfo
webdataset
opencv-python